App Settings, Remote Server Connection & Config, and Maven Jar Package
Settings can be taken in from either a user specified file/location or from $home/.config/BackupValidation.  Can also create a config template.

Remote Connection and Remote Server Configuration are working.

When there is a main function in team3.App the maven package phase should work to create a jar that runs.

Working on an entry point with CLI options in team3.App but it's not ready to commit yet.